third_party/blink/web_tests/external/wpt/websockets/interfaces/WebSocket/readyState/003-expected.txt

This is a testharness.js-based test.
FAIL WebSockets: delete readyState assert_equals: delete ws.readyState expected 2 but got 3
Harness: the test ran to completion.
